Extract Prompt Studio text helpers
